feat: regenerate on each reload
test update + each call will regenerate the food
